[redhat] New configs in drivers/mfd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MFD_INTEL_M10_BMC_SPI [=n]
 Type  : tristate
 Defined at drivers/mfd/Kconfig:2176
   Prompt: Intel MAX 10 Board Management Controller with SPI
   Depends on: HAS_IOMEM [=y] && SPI_MASTER [=y]
   Location:
     -> Device Drivers
       -> Multifunction device drivers
         -> Intel MAX 10 Board Management Controller with SPI (MFD_INTEL_M10_BMC_SPI [=n])
 Selects: MFD_INTEL_M10_BMC_CORE [=n] && REGMAP_SPI_AVMM [=n]

---

Cc: Mark Salter <msalter@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_MFD_INTEL_M10_BMC_SPI enable as a module
v3:
- CONFIG_SENSORS_INTEL_M10_BMC_HWMON enabled as a module

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>