redhat: configs: Enable CONFIG_MEMTEST to enable memory test
The ability to quickly check for memory problems is invaluable. A bad DIMM
chip or some other issue will present in awkward ways. The ability to turn
on this check and then automatically mark bad sectors of RAM as buffer is
very useful.

The hardware tested has physical water damage issues, but only with certain
addresses. Utilizing this option helped to diagnose the issue.

Link: https://bugzilla.redhat.com/show_bug.cgi?id=2182993
Signed-off-by: Kate Hsuan <hpa@redhat.com>