redhat/Makefile: Support building linux-next
Add support for building linux-next with code similar to that used for
building mainline snapshots. The code to adjust SPECKEXTRAVERSION
during the merge window is copied from there. To ensure the upgrade
path start the paclage release with "0.0"; also add the tag "next"
together with the release date to make it clear which version of next
is in the package.

Signed-off-by: Thorsten Leemhuis <fedora@leemhuis.info>