[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: KUNPENG_HCCS [=n]
 Type  : tristate
 Defined at drivers/soc/hisilicon/Kconfig:6
   Prompt: HCCS driver on Kunpeng SoC
   Depends on: (ARCH_HISI [=y] || COMPILE_TEST [=n]) && ACPI [=y] && MAILBOX [=y] && (ARM64 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Hisilicon SoC drivers
           -> HCCS driver on Kunpeng SoC (KUNPENG_HCCS [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>