Don't use upstream bpftool version for Fedora package
Upstream bpftool went with their own versioning. While this works in
rawhide and RHEL where release is bumped wth every build, it does not
work with Fedora. Because bpftool sublevel is not bumped with stable
updates, and we don't update the release in Fedora when the version has
changed, we ended up with a failure because the bpftool subpackage
version already exists. For Fedora, we will stick with the kernel
versioning for the package. bpftool -V still displays the upstream
bpftool version number.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>