Automation cleanups for rebasing rt-devel and automotive-devel
The -rt tree updated to 6.9-rc2 and forced a rebased on the rt-devel
branch.  The rebase logic had some quirks in it.

* missing merge message
* make previous branches easier to use
* filter out Automatic merges as redundant with new base
* allow force pushes on rebase only

These changes allow the rebase from 6.8 to 6.9 to happen successfully.

The sed script will need to be revisited later as it is fragile.
Hopefully rt can merge sooner than later so we can drop the rebase
stuff.

Signed-off-by: Don Zickus <dzickus@redhat.com>