include/linux: Adjust headers for C23
GCC 15 changed default from -std=gnu17 to -std=gnu23.  In C23
among many other changes bool, false and true are keywords
(like in C++), so defining those using typedef or enum is invalid.

The following patch adjusts the include/linux/ headers to be C23
compatible.  _Bool and the C23 bool are ABI compatible, false/true
have the same values but different types (previously in the kernel
case it was an anonymous enum, in C23 it is bool), so if something uses
say sizeof(false) or typeof(true), those do change, but I doubt that
is used anywhere in the kernel.

The last change is about va_start.  In C23 ellipsis can be specified
without any arguments before it, like
int foo(...)
{
	va_list ap;
	va_start(ap);
	int ret = va_arg(ap, int);
	va_end(ap);
	return ret;
}
and unlike in C17 and earlier, va_start is macro with variable argument
count.  Normally one should use it with just one argument or for backwards
compatibility with C17 and earlier with two arguments, second being the last
named argument.  Of course, if there is no last named argument, only the
single argument va_start is an option.  The stdarg.h change attempts to be
compatible with older versions of GCC and with clang as well.  Both GCC 13-14
and recent versions of clang define va_start for C23 as
#define va_start(v, ...) __builtin_va_start(v, 0)
The problem with that definition is that it doesn't emit warnings when one
uses complete nonsense in there (e.g. va_start(ap, 8) or
va_start(ap, +-*, /, 3, 4.0)), so for GCC 15 it uses a different builtin
which takes care about warnings for unexpected va_start uses (as suggested
by the C23 standard).  Hopefully clang will one day implement that too.

Anyway, without these changes, kernel could detect compiler defaulting to
C23 and use say -std=gnu17 option instead, but even in that case IMHO this
patch doesn't hurt.

Signed-off-by: Jakub Jelinek <jakub@redhat.com>