not upstream: Set vdso is_ready later during boot process
Instead of setting vdso is_ready when the CRNG becomes ready, do
it later in the boot process as FIPS mode itself may not have been
set yet during the early-boot stages.

In particular, the CRNG gets initialised very early on arm64, even
before the early boot paramaters are parsed.

Delaying vdso is_ready setting on the other hand is harmless as
there is always a fallback code path that leads to getrandom(2).

Signed-off-by: Herbert Xu <herbert.xu@redhat.com>