kernel.spec: update sources for automotive-only build
When we generate the kernel-automotive srpm, we only want to include
sources that will be required for that build. For automotive the
INCLUDE_FEDORA_FILES, INCLUDE_RHEL_FILES and INCLUDE_RT_FILES will be
disabled which means we need to arrage the sources so that
kernel-automotive will have all the files it requires.

Signed-off-by: Scott Weaver <scweaver@redhat.com>